........
  r1115394 | ganeshp | 2010-04-16 12:03:56 +0200 (Fri, 16 Apr 2010) | 1 line
  
  The Fix for the crash problem pointer out by Thorsten ( Please see commit log for the previous version of this file for details on the bug )
........
  r1115499 | zander | 2010-04-16 18:07:03 +0200 (Fri, 16 Apr 2010) | 1 line
  
  Disable unfinished feature (we can't save it) 'page spreads'
........
  r1115500 | zander | 2010-04-16 18:07:06 +0200 (Fri, 16 Apr 2010) | 1 line
  
  Remove some debug
........
  r1115501 | zander | 2010-04-16 18:07:09 +0200 (Fri, 16 Apr 2010) | 1 line
  
  Add product name
........
  r1115502 | zander | 2010-04-16 18:07:11 +0200 (Fri, 16 Apr 2010) | 1 line
  
  Fix API docs errors
........
  r1115503 | zander | 2010-04-16 18:07:13 +0200 (Fri, 16 Apr 2010) | 1 line
  
  Fix various API docs errors
........
  r1115504 | zander | 2010-04-16 18:07:15 +0200 (Fri, 16 Apr 2010) | 4 lines
  
  Add virtual keyword to make this clear
  
  In C++ a virtual method is always virtual, also in all inheriting
  classes. To make this clear to the reader its good to repeat that.
........
  r1115505 | zander | 2010-04-16 18:07:17 +0200 (Fri, 16 Apr 2010) | 1 line
  
  Fix valgrind issue; don't access empty array
........
  r1115542 | piggz | 2010-04-16 21:46:03 +0200 (Fri, 16 Apr 2010) | 4 lines
  
  Use qint64 instead of long for record position/count.  Add basic position support for the migration driver so that it can render properly across pages.
  CCMAIL:danders@get2net.dk
  CCMAIL:staniek@kde.org
........
  r1115720 | dkazakov | 2010-04-17 11:15:07 +0200 (Sat, 17 Apr 2010) | 7 lines
  
  Fixed an Q_ASSERT(!m_numTiles) bug
  
  There was a race condition in a copy-constructor of the hash
  table. The source table should have been locked first to get a safe
  clone.
  BUG:232524
........
  r1115725 | rempt | 2010-04-17 11:43:35 +0200 (Sat, 17 Apr 2010) | 1 line
  
  warnings--
........
  r1115726 | rempt | 2010-04-17 11:43:39 +0200 (Sat, 17 Apr 2010) | 1 line
  
  warnings--
........
  r1115727 | rempt | 2010-04-17 11:43:43 +0200 (Sat, 17 Apr 2010) | 4 lines
  
  warning--
  
  Four years after porting started, it's unlikely anyone will
  reconstruct what this was for and port it.
........
  r1115728 | rempt | 2010-04-17 11:43:52 +0200 (Sat, 17 Apr 2010) | 1 line
  
  warning--
........
  r1115729 | rempt | 2010-04-17 11:43:56 +0200 (Sat, 17 Apr 2010) | 1 line
  
  warnings--
........
  r1115730 | rempt | 2010-04-17 11:44:01 +0200 (Sat, 17 Apr 2010) | 1 line
  
  warnings--
........
  r1115731 | rempt | 2010-04-17 11:44:04 +0200 (Sat, 17 Apr 2010) | 1 line
  
  warnings--
........
  r1115732 | rempt | 2010-04-17 11:44:07 +0200 (Sat, 17 Apr 2010) | 1 line
  
  warnings--
........
  r1115733 | rempt | 2010-04-17 11:44:11 +0200 (Sat, 17 Apr 2010) | 1 line
  
  warnings--
........
  r1115734 | rempt | 2010-04-17 11:44:18 +0200 (Sat, 17 Apr 2010) | 1 line
  
  warnings--
........
  r1115735 | rempt | 2010-04-17 11:44:22 +0200 (Sat, 17 Apr 2010) | 4 lines
  
  Undeprecate createDomDocument
  
  This method will always be necessary for supporting Krita's native
  file format, so it should not be deprecated.
........
  r1115736 | rempt | 2010-04-17 11:44:25 +0200 (Sat, 17 Apr 2010) | 1 line
  
  warning--
........
  r1115737 | rempt | 2010-04-17 11:44:35 +0200 (Sat, 17 Apr 2010) | 1 line
  
  warnings--
........
  r1115738 | rempt | 2010-04-17 11:44:39 +0200 (Sat, 17 Apr 2010) | 1 line
  
  warnings--
........
  r1115739 | rempt | 2010-04-17 11:44:42 +0200 (Sat, 17 Apr 2010) | 1 line
  
  warnings--
........
  r1115740 | rempt | 2010-04-17 11:44:44 +0200 (Sat, 17 Apr 2010) | 1 line
  
  warning--
........
  r1115741 | rempt | 2010-04-17 11:44:47 +0200 (Sat, 17 Apr 2010) | 1 line
  
  Fix comment
........

svn path=/branches/work/koffice-ko/; revision=1115743

+++ krita/plugins/colorspaces/ctlcs/KoCtlColorTransformationFactory.cpp
@@ -134,7 +134,7 @@ KoColorTransformation* KoCtlColorTransformationFactory::createTransformation(con
     GTLCore::PixelDescription pixelDescription = createPixelDescription(colorSpace);
     dbgPlugins << pixelDescription.bitsSize() << " " << colorSpace->pixelSize();
 
-    Q_ASSERT(pixelDescription.bitsSize() / 8 == colorSpace->pixelSize());
+    Q_ASSERT(pixelDescription.bitsSize() / 8 == (int)colorSpace->pixelSize());
 
     QMutexLocker lock2(&m_mutex);
