+++ krita/plugins/colorspaces/ctlcs/KoCtlMixColorsOp.cpp
@@ -51,7 +51,7 @@ void KoCtlMixColorsOp::mixColors(const quint8 * const* colors, const qint16 *wei
         alphaTimesWeight *= *weights;
 
         for (uint i = 0; i < m_colorSpace->channelCount(); i++) {
-            if (i != alphaPos) {
+            if ((int)i != alphaPos) {
                 m_accumulators[i]->mix(color, alphaTimesWeight);
             }
         }
@@ -65,7 +65,7 @@ void KoCtlMixColorsOp::mixColors(const quint8 * const* colors, const qint16 *wei
     if (totalAlpha > 0) {
         double inv = 1.0 / (totalAlpha * 255);
         for (uint i = 0; i < m_colorSpace->channelCount(); i++) {
-            if (i != alphaPos) {
+            if ((int)i != alphaPos) {
                 m_accumulators[i]->affect(dst, inv);
             }
         }
